backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9

# Azure OpenAI
openai==1.3.5

# PDF Processing (minimal)
PyPDF2==3.0.1

# Azure Storage
azure-storage-blob==12.19.0

# Environment
python-dotenv==1.0.0

# CORS
fastapi-cors==0.1.0

# MINIMAL LangChain components ONLY
langchain-core==0.1.0
langchain-community==0.0.10
langchain-text-splitters==0.0.1

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0